fix: move utils to avoid sharing code between e2e and app
