+++ identity/client/e2e/utils/relativizePaths.ts
